long -> int

+++ llm-service/src/main/java/de/flowsuite/llmservice/common/ModelResponse.java
@@ -1,4 +1,4 @@
 package de.flowsuite.llmservice.common;
 
 public record ModelResponse(
-        String text, String modelName, long inputTokens, long outputTokens, long totalTokens) {}
+        String text, String modelName, int inputTokens, int outputTokens, int totalTokens) {}
